extract exercise deletion service to simplify dependency management
