+++ src/main/java/de/tum/in/www1/artemis/service/CourseService.java
@@ -50,6 +50,8 @@ public class CourseService {
 
     private final ExerciseService exerciseService;
 
+    private final ExerciseDeletionService exerciseDeletionService;
+
     private final AuthorizationCheckService authCheckService;
 
     private final LectureService lectureService;
@@ -78,12 +80,14 @@ public class CourseService {
 
     private final GradingScaleRepository gradingScaleRepository;
 
-    public CourseService(CourseRepository courseRepository, ExerciseService exerciseService, AuthorizationCheckService authCheckService, UserRepository userRepository,
-            LectureService lectureService, GroupNotificationRepository groupNotificationRepository, ExerciseGroupRepository exerciseGroupRepository,
-            AuditEventRepository auditEventRepository, UserService userService, LearningGoalRepository learningGoalRepository, GroupNotificationService groupNotificationService,
-            ExamService examService, ExamRepository examRepository, CourseExamExportService courseExamExportService, GradingScaleRepository gradingScaleRepository) {
+    public CourseService(CourseRepository courseRepository, ExerciseService exerciseService, ExerciseDeletionService exerciseDeletionService,
+            AuthorizationCheckService authCheckService, UserRepository userRepository, LectureService lectureService, GroupNotificationRepository groupNotificationRepository,
+            ExerciseGroupRepository exerciseGroupRepository, AuditEventRepository auditEventRepository, UserService userService, LearningGoalRepository learningGoalRepository,
+            GroupNotificationService groupNotificationService, ExamService examService, ExamRepository examRepository, CourseExamExportService courseExamExportService,
+            GradingScaleRepository gradingScaleRepository) {
         this.courseRepository = courseRepository;
         this.exerciseService = exerciseService;
+        this.exerciseDeletionService = exerciseDeletionService;
         this.authCheckService = authCheckService;
         this.userRepository = userRepository;
         this.lectureService = lectureService;
@@ -168,7 +172,7 @@ public class CourseService {
      * <ul>
      *     <li>The Course</li>
      *     <li>All Exercises including:
-     *      submissions, participations, results, repositories and build plans, see {@link ExerciseService#delete}</li>
+     *      submissions, participations, results, repositories and build plans, see {@link ExerciseDeletionService#delete}</li>
      *     <li>All Lectures and their Attachments, see {@link LectureService#delete}</li>
      *     <li>All GroupNotifications of the course, see {@link GroupNotificationRepository#delete}</li>
      *     <li>All default groups created by Artemis, see {@link UserService#deleteGroup}</li>
@@ -236,7 +240,7 @@ public class CourseService {
 
     private void deleteExercisesOfCourse(Course course) {
         for (Exercise exercise : course.getExercises()) {
-            exerciseService.delete(exercise.getId(), true, true);
+            exerciseDeletionService.delete(exercise.getId(), true, true);
         }
     }
 
@@ -559,7 +563,7 @@ public class CourseService {
         var exercisesToCleanup = Stream.concat(course.getExercises().stream(), examExercises.stream()).collect(Collectors.toSet());
         exercisesToCleanup.forEach(exercise -> {
             if (exercise instanceof ProgrammingExercise) {
-                exerciseService.cleanup(exercise.getId(), true);
+                exerciseDeletionService.cleanup(exercise.getId(), true);
             }
 
             // TODO: extend exerciseService.cleanup to clean up all exercise types
