+++ src/main/java/de/tum/in/www1/artemis/service/ExerciseService.java
@@ -8,29 +8,21 @@ import java.util.stream.Collectors;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.actuate.audit.AuditEvent;
 import org.springframework.boot.actuate.audit.AuditEventRepository;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import de.tum.in.www1.artemis.config.Constants;
 import de.tum.in.www1.artemis.domain.*;
 import de.tum.in.www1.artemis.domain.enumeration.ComplaintType;
 import de.tum.in.www1.artemis.domain.enumeration.ExerciseMode;
 import de.tum.in.www1.artemis.domain.enumeration.InitializationState;
-import de.tum.in.www1.artemis.domain.exam.Exam;
-import de.tum.in.www1.artemis.domain.exam.StudentExam;
-import de.tum.in.www1.artemis.domain.lecture.ExerciseUnit;
-import de.tum.in.www1.artemis.domain.modeling.ModelingExercise;
-import de.tum.in.www1.artemis.domain.participation.ProgrammingExerciseStudentParticipation;
 import de.tum.in.www1.artemis.domain.participation.StudentParticipation;
 import de.tum.in.www1.artemis.domain.quiz.QuizExercise;
 import de.tum.in.www1.artemis.domain.quiz.QuizSubmission;
 import de.tum.in.www1.artemis.domain.scores.ParticipantScore;
 import de.tum.in.www1.artemis.repository.*;
 import de.tum.in.www1.artemis.service.programming.ProgrammingAssessmentService;
-import de.tum.in.www1.artemis.service.programming.ProgrammingExerciseService;
 import de.tum.in.www1.artemis.service.scheduled.quiz.QuizScheduleService;
 import de.tum.in.www1.artemis.web.rest.dto.CourseManagementOverviewExerciseStatisticsDTO;
 import de.tum.in.www1.artemis.web.rest.dto.DueDateStat;
@@ -45,40 +37,19 @@ import de.tum.in.www1.artemis.web.rest.errors.EntityNotFoundException;
 @Service
 public class ExerciseService {
 
-    @Value("${jhipster.clientApp.name}")
-    private String applicationName;
-
     private final Logger log = LoggerFactory.getLogger(ExerciseService.class);
-
-    private final ParticipationService participationService;
 
     private final AuthorizationCheckService authCheckService;
 
-    private final ProgrammingExerciseService programmingExerciseService;
-
-    private final ModelingExerciseService modelingExerciseService;
-
-    private final QuizExerciseService quizExerciseService;
-
     private final QuizScheduleService quizScheduleService;
-
-    private final ExampleSubmissionService exampleSubmissionService;
 
     private final ProgrammingAssessmentService programmingAssessmentService;
 
     private final TeamRepository teamRepository;
 
-    private final ExamRepository examRepository;
-
-    private final StudentExamRepository studentExamRepository;
-
     private final AuditEventRepository auditEventRepository;
 
-    private final ExerciseUnitRepository exerciseUnitRepository;
-
     private final ExerciseRepository exerciseRepository;
-
-    private final TutorParticipationRepository tutorParticipationRepository;
 
     private final ParticipantScoreRepository participantScoreRepository;
 
@@ -89,8 +60,6 @@ public class ExerciseService {
     private final LtiOutcomeUrlRepository ltiOutcomeUrlRepository;
 
     private final StudentParticipationRepository studentParticipationRepository;
-
-    private final LectureUnitService lectureUnitService;
 
     private final UserRepository userRepository;
 
@@ -106,50 +75,25 @@ public class ExerciseService {
 
     private final FeedbackRepository feedbackRepository;
 
-    private final PlagiarismResultRepository plagiarismResultRepository;
-
     private final RatingService ratingService;
 
-    private final TextAssessmentKnowledgeService textAssessmentKnowledgeService;
-
-    private final ModelAssessmentKnowledgeService modelAssessmentKnowledgeService;
-
-    private final TextExerciseRepository textExerciseRepository;
-
-    private final ModelingExerciseRepository modelingExerciseRepository;
-
-    public ExerciseService(ExerciseRepository exerciseRepository, ExerciseUnitRepository exerciseUnitRepository, ParticipationService participationService,
-            AuthorizationCheckService authCheckService, ProgrammingExerciseService programmingExerciseService, ModelingExerciseService modelingExerciseService,
-            QuizExerciseService quizExerciseService, QuizScheduleService quizScheduleService, TutorParticipationRepository tutorParticipationRepository,
-            ExampleSubmissionService exampleSubmissionService, AuditEventRepository auditEventRepository, TeamRepository teamRepository,
-            StudentExamRepository studentExamRepository, ExamRepository examRepository, ProgrammingExerciseRepository programmingExerciseRepository,
+    public ExerciseService(ExerciseRepository exerciseRepository, AuthorizationCheckService authCheckService, QuizScheduleService quizScheduleService,
+            AuditEventRepository auditEventRepository, TeamRepository teamRepository, ProgrammingExerciseRepository programmingExerciseRepository,
             LtiOutcomeUrlRepository ltiOutcomeUrlRepository, StudentParticipationRepository studentParticipationRepository, ResultRepository resultRepository,
-            SubmissionRepository submissionRepository, ParticipantScoreRepository participantScoreRepository, LectureUnitService lectureUnitService, UserRepository userRepository,
+            SubmissionRepository submissionRepository, ParticipantScoreRepository participantScoreRepository, UserRepository userRepository,
             ComplaintRepository complaintRepository, TutorLeaderboardService tutorLeaderboardService, ComplaintResponseRepository complaintResponseRepository,
-            PlagiarismResultRepository plagiarismResultRepository, GradingCriterionRepository gradingCriterionRepository, FeedbackRepository feedbackRepository,
-            ProgrammingAssessmentService programmingAssessmentService, TextAssessmentKnowledgeService textAssessmentKnowledgeService,
-            ModelAssessmentKnowledgeService modelAssessmentKnowledgeService, TextExerciseRepository textExerciseRepository, ModelingExerciseRepository modelingExerciseRepository,
+            GradingCriterionRepository gradingCriterionRepository, FeedbackRepository feedbackRepository, ProgrammingAssessmentService programmingAssessmentService,
             RatingService ratingService) {
         this.exerciseRepository = exerciseRepository;
         this.resultRepository = resultRepository;
-        this.examRepository = examRepository;
-        this.participationService = participationService;
         this.authCheckService = authCheckService;
-        this.programmingExerciseService = programmingExerciseService;
-        this.modelingExerciseService = modelingExerciseService;
-        this.tutorParticipationRepository = tutorParticipationRepository;
-        this.exampleSubmissionService = exampleSubmissionService;
         this.auditEventRepository = auditEventRepository;
-        this.quizExerciseService = quizExerciseService;
         this.quizScheduleService = quizScheduleService;
-        this.studentExamRepository = studentExamRepository;
-        this.exerciseUnitRepository = exerciseUnitRepository;
         this.submissionRepository = submissionRepository;
         this.teamRepository = teamRepository;
         this.participantScoreRepository = participantScoreRepository;
         this.ltiOutcomeUrlRepository = ltiOutcomeUrlRepository;
         this.studentParticipationRepository = studentParticipationRepository;
-        this.lectureUnitService = lectureUnitService;
         this.userRepository = userRepository;
         this.complaintRepository = complaintRepository;
         this.tutorLeaderboardService = tutorLeaderboardService;
@@ -158,12 +102,7 @@ public class ExerciseService {
         this.gradingCriterionRepository = gradingCriterionRepository;
         this.feedbackRepository = feedbackRepository;
         this.programmingAssessmentService = programmingAssessmentService;
-        this.plagiarismResultRepository = plagiarismResultRepository;
         this.ratingService = ratingService;
-        this.textAssessmentKnowledgeService = textAssessmentKnowledgeService;
-        this.modelAssessmentKnowledgeService = modelAssessmentKnowledgeService;
-        this.textExerciseRepository = textExerciseRepository;
-        this.modelingExerciseRepository = modelingExerciseRepository;
     }
 
     /**
@@ -374,116 +313,6 @@ public class ExerciseService {
     }
 
     /**
-     * Checks if the exercise has any test runs and sets the transient property if it does
-     * @param exercise - the exercise for which we check if test runs exist
-     */
-    public void checkTestRunsExist(Exercise exercise) {
-        Long containsTestRunParticipations = studentParticipationRepository.countParticipationsOnlyTestRunsByExerciseId(exercise.getId());
-        if (containsTestRunParticipations != null && containsTestRunParticipations > 0) {
-            exercise.setTestRunParticipationsExist(Boolean.TRUE);
-        }
-    }
-
-    /**
-     * Resets an Exercise by deleting all its participations and plagiarsim results
-     *
-     * @param exercise which should be reset
-     */
-    public void reset(Exercise exercise) {
-        log.debug("Request reset Exercise : {}", exercise.getId());
-
-        // delete all plagiarism results for this exercise
-        plagiarismResultRepository.deletePlagiarismResultsByExerciseId(exercise.getId());
-
-        // delete all participations for this exercise
-        participationService.deleteAllByExerciseId(exercise.getId(), true, true);
-
-        if (exercise instanceof QuizExercise) {
-            quizExerciseService.resetExercise(exercise.getId());
-        }
-    }
-
-    /**
-     * Delete the exercise by id and all its participations.
-     *
-     * @param exerciseId                   the exercise to be deleted
-     * @param deleteStudentReposBuildPlans whether the student repos and build plans should be deleted (can be true for programming exercises and should be false for all other exercise types)
-     * @param deleteBaseReposBuildPlans    whether the template and solution repos and build plans should be deleted (can be true for programming exercises and should be false for all other exercise types)
-     */
-    @Transactional // ok
-    public void delete(long exerciseId, boolean deleteStudentReposBuildPlans, boolean deleteBaseReposBuildPlans) {
-        // Delete has a transactional mechanism. Therefore, all lazy objects that are deleted below, should be fetched when needed.
-        final var exercise = exerciseRepository.findByIdElseThrow(exerciseId);
-
-        log.info("Checking if exercise {} is modeling exercise", exercise.getId());
-        if (exercise instanceof ModelingExercise) {
-            log.info("Deleting clusters, elements and cancel scheduled operations of exercise {}", exercise.getId());
-
-            modelingExerciseService.deleteClustersAndElements((ModelingExercise) exercise);
-            modelingExerciseService.cancelScheduledOperations(exerciseId);
-        }
-
-        participantScoreRepository.deleteAllByExerciseIdTransactional(exerciseId);
-        // delete all exercise units linking to the exercise
-        List<ExerciseUnit> exerciseUnits = this.exerciseUnitRepository.findByIdWithLearningGoalsBidirectional(exerciseId);
-        for (ExerciseUnit exerciseUnit : exerciseUnits) {
-            this.lectureUnitService.removeLectureUnit(exerciseUnit);
-        }
-
-        // delete all plagiarism results belonging to this exercise
-        plagiarismResultRepository.deletePlagiarismResultsByExerciseId(exerciseId);
-
-        // delete all participations belonging to this quiz
-        participationService.deleteAllByExerciseId(exercise.getId(), deleteStudentReposBuildPlans, deleteStudentReposBuildPlans);
-        // clean up the many-to-many relationship to avoid problems when deleting the entities but not the relationship table
-        // to avoid a ConcurrentModificationException, we need to use a copy of the set
-        var exampleSubmissions = new HashSet<>(exercise.getExampleSubmissions());
-        for (ExampleSubmission exampleSubmission : exampleSubmissions) {
-            exampleSubmissionService.deleteById(exampleSubmission.getId());
-        }
-        // make sure tutor participations are deleted before the exercise is deleted
-        tutorParticipationRepository.deleteAllByAssessedExerciseId(exercise.getId());
-
-        if (exercise.isExamExercise()) {
-            Exam exam = examRepository.findOneWithEagerExercisesGroupsAndStudentExams(exercise.getExerciseGroup().getExam().getId());
-            for (StudentExam studentExam : exam.getStudentExams()) {
-                if (studentExam.getExercises().contains(exercise)) {
-                    // remove exercise reference from student exam
-                    List<Exercise> exerciseList = studentExam.getExercises();
-                    exerciseList.remove(exercise);
-                    studentExam.setExercises(exerciseList);
-                    studentExamRepository.save(studentExam);
-                }
-            }
-        }
-
-        // Programming exercises have some special stuff that needs to be cleaned up (solution/template participation, build plans, etc.).
-        if (exercise instanceof ProgrammingExercise) {
-            // TODO: delete all schedules related to this programming exercise
-            programmingExerciseService.delete(exercise.getId(), deleteBaseReposBuildPlans);
-        }
-        else {
-            // delete text assessment knowledge if exercise is of type TextExercise and if no other exercise uses same knowledge
-            if (exercise instanceof TextExercise) {
-                // explicitly load the text exercise as such so that the knowledge is eagerly loaded as well
-                TextExercise textExercise = textExerciseRepository.findByIdElseThrow(exercise.getId());
-                if (textExercise.getKnowledge() != null) {
-                    textAssessmentKnowledgeService.deleteKnowledge(textExercise.getKnowledge().getId(), textExercise.getId());
-                }
-            }
-            // delete model assessment knowledge if exercise is of type ModelExercise and if no other exercise uses same knowledge
-            else if (exercise instanceof ModelingExercise) {
-                // explicitly load the modeling exercise as such so that the knowledge is eagerly loaded as well
-                ModelingExercise modelingExercise = modelingExerciseRepository.findByIdElseThrow(exercise.getId());
-                if (modelingExercise.getKnowledge() != null) {
-                    modelAssessmentKnowledgeService.deleteKnowledge(modelingExercise.getKnowledge().getId(), modelingExercise.getId());
-                }
-            }
-            exerciseRepository.delete(exercise);
-        }
-    }
-
-    /**
      * Updates the points of related exercises if the points of exercises have changed
      *
      * @param originalExercise the original exercise
@@ -510,35 +339,6 @@ public class ExerciseService {
             participantScore.setLastRatedPoints(lastRatedPoints);
         }
         participantScoreRepository.saveAll(participantScoreList);
-    }
-
-    /**
-     * Delete student build plans (except BASE/SOLUTION) and optionally git repositories of all exercise student participations.
-     *
-     * @param exerciseId         programming exercise for which build plans in respective student participations are deleted
-     * @param deleteRepositories if true, the repositories gets deleted
-     */
-    public void cleanup(Long exerciseId, boolean deleteRepositories) {
-        Exercise exercise = exerciseRepository.findByIdWithStudentParticipationsElseThrow(exerciseId);
-        log.info("Request to cleanup all participations for Exercise : {}", exercise.getTitle());
-
-        if (exercise instanceof ProgrammingExercise) {
-            for (StudentParticipation participation : exercise.getStudentParticipations()) {
-                participationService.cleanupBuildPlan((ProgrammingExerciseStudentParticipation) participation);
-            }
-
-            if (!deleteRepositories) {
-                return;    // in this case, we are done
-            }
-
-            for (StudentParticipation participation : exercise.getStudentParticipations()) {
-                participationService.cleanupRepository((ProgrammingExerciseStudentParticipation) participation);
-            }
-
-        }
-        else {
-            log.warn("Exercise with exerciseId {} is not an instance of ProgrammingExercise. Ignoring the request to cleanup repositories and build plan", exerciseId);
-        }
     }
 
     public void logDeletion(Exercise exercise, Course course, User user) {
@@ -742,14 +542,14 @@ public class ExerciseService {
      * @param exercise the exercise corresponding to the <code>CourseManagementOverviewExerciseStatisticsDTO</code>
      */
     private void setAverageScoreForStatisticsDTO(CourseManagementOverviewExerciseStatisticsDTO exerciseStatisticsDTO, Map<Long, Double> averageScoreById, Exercise exercise) {
+        Double averageScore;
         if (exercise instanceof QuizExercise) {
-            var averageScore = participantScoreRepository.findAverageScoreForExercise(exercise.getId());
-            exerciseStatisticsDTO.setAverageScoreInPercent(averageScore != null ? averageScore : 0.0);
+            averageScore = participantScoreRepository.findAverageScoreForExercise(exercise.getId());
         }
         else {
-            var averageScore = averageScoreById.get(exercise.getId());
-            exerciseStatisticsDTO.setAverageScoreInPercent(averageScore != null ? averageScore : 0.0);
+            averageScore = averageScoreById.get(exercise.getId());
         }
+        exerciseStatisticsDTO.setAverageScoreInPercent(averageScore != null ? averageScore : 0.0);
     }
 
     /**
@@ -945,7 +745,7 @@ public class ExerciseService {
                     .map(GradingInstruction::getId).toList();
 
             // collect deleted grading instruction ids into the list
-            List<Long> gradingInstructionIdsToBeDeleted = backupInstructionIds.stream().filter(backupinstructionId -> !updatedInstructionIds.contains(backupinstructionId))
+            List<Long> gradingInstructionIdsToBeDeleted = backupInstructionIds.stream().filter(backupInstructionId -> !updatedInstructionIds.contains(backupInstructionId))
                     .toList();
 
             // determine the feedback to be deleted
