+++ src/main/java/de/tum/in/www1/artemis/web/rest/QuizExerciseResource.java
@@ -22,6 +22,7 @@ import de.tum.in.www1.artemis.domain.User;
 import de.tum.in.www1.artemis.domain.quiz.QuizExercise;
 import de.tum.in.www1.artemis.repository.CourseRepository;
 import de.tum.in.www1.artemis.repository.QuizExerciseRepository;
+import de.tum.in.www1.artemis.repository.StudentParticipationRepository;
 import de.tum.in.www1.artemis.repository.UserRepository;
 import de.tum.in.www1.artemis.security.Role;
 import de.tum.in.www1.artemis.service.*;
@@ -58,6 +59,8 @@ public class QuizExerciseResource {
 
     private final ExerciseService exerciseService;
 
+    private final ExerciseDeletionService exerciseDeletionService;
+
     private final ExamDateService examDateService;
 
     private final QuizScheduleService quizScheduleService;
@@ -70,12 +73,16 @@ public class QuizExerciseResource {
 
     private final InstanceMessageSendService instanceMessageSendService;
 
+    private final StudentParticipationRepository studentParticipationRepository;
+
     public QuizExerciseResource(QuizExerciseService quizExerciseService, QuizExerciseRepository quizExerciseRepository, CourseService courseService,
-            QuizScheduleService quizScheduleService, QuizStatisticService quizStatisticService, AuthorizationCheckService authCheckService, CourseRepository courseRepository,
-            GroupNotificationService groupNotificationService, ExerciseService exerciseService, UserRepository userRepository, ExamDateService examDateService,
-            QuizMessagingService quizMessagingService, InstanceMessageSendService instanceMessageSendService) {
+            ExerciseDeletionService exerciseDeletionServiceService, QuizScheduleService quizScheduleService, QuizStatisticService quizStatisticService,
+            AuthorizationCheckService authCheckService, CourseRepository courseRepository, GroupNotificationService groupNotificationService, ExerciseService exerciseService,
+            UserRepository userRepository, ExamDateService examDateService, QuizMessagingService quizMessagingService, InstanceMessageSendService instanceMessageSendService,
+            StudentParticipationRepository studentParticipationRepository) {
         this.quizExerciseService = quizExerciseService;
         this.quizExerciseRepository = quizExerciseRepository;
+        this.exerciseDeletionService = exerciseDeletionServiceService;
         this.userRepository = userRepository;
         this.courseService = courseService;
         this.quizScheduleService = quizScheduleService;
@@ -87,6 +94,7 @@ public class QuizExerciseResource {
         this.courseRepository = courseRepository;
         this.quizMessagingService = quizMessagingService;
         this.instanceMessageSendService = instanceMessageSendService;
+        this.studentParticipationRepository = studentParticipationRepository;
     }
 
     /**
@@ -261,7 +269,7 @@ public class QuizExerciseResource {
             if (!authCheckService.isAtLeastEditorInCourse(course, null)) {
                 return forbidden();
             }
-            exerciseService.checkTestRunsExist(quizExercise);
+            studentParticipationRepository.checkTestRunsExist(quizExercise);
         }
         else if (!authCheckService.isAllowedToSeeExercise(quizExercise, null)) {
             return forbidden();
@@ -405,7 +413,7 @@ public class QuizExerciseResource {
         }
         // note: we use the exercise service here, because this one makes sure to clean up all lazy references correctly.
         exerciseService.logDeletion(quizExercise, course, user);
-        exerciseService.delete(quizExerciseId, false, false);
+        exerciseDeletionService.delete(quizExerciseId, false, false);
         quizExerciseService.cancelScheduledQuiz(quizExerciseId);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, quizExercise.getTitle())).build();
     }
